aws/sam/mads-simulate-app/swap_simulation/requirements.txt
requests
aws-psycopg2
pandas
pytz
sklearn
-f https://download.pytorch.org/whl/torch_stable.html 
torch==1.9.0+cpu 
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.10.0+cpu